Add files via upload
Health Insurance Marketplace:
The data used in this project is taken from kaggle.com. It was originally prepared and released by the Centers for Medicare & Medicaid Services (CMS). The goal of this project is to get familiar with the Python basic functions and to get exposure to a full life cycle of a data project.